Move some classes to base lib for reuse
