+++ FreeMote.Psb/FreeMote.Psb.csproj
@@ -32,6 +32,7 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Drawing" />
     <Reference Include="System.Runtime" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -44,8 +45,10 @@
     <Compile Include="PSB.cs" />
     <Compile Include="PsbFileExtension.cs" />
     <Compile Include="PsbHelper.cs" />
+    <Compile Include="PsbResCollector.cs" />
     <Compile Include="PsbTypes.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ResourceMetadata.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\FreeMote\FreeMote.csproj">
